Splitting afni tools to own py folder
