--- run
@@ -1,15 +1,0 @@
-#!/bin/bash
-source /etc/afni/afni.sh
-Xvfb :88 -screen 0 1024x768x24 >& /dev/null &
-export DISPLAY=:88
-export PYTHON_PATH=$PYTHON_PATH:/flywheel/v0/me-ica/
-python /flywheel/v0/run_meica.py
-exit_status=$?
-killall Xvfb
-if [[ $exit_status == "0" ]] && [[ -n $(find /flywheel/v0/output/* -name "*medn*.nii.gz") ]]; then
-  echo -e "MEICA success! (exit code = $exit_status)"
-  exit 0
-else
-  echo -e "Errors detected (exit code = $exit_status)"
-  exit 1
-fi
