+++ run_meica.py
@@ -7,51 +7,15 @@ import logging
 import datetime
 import os
 import subprocess as sp
-from collections import OrderedDict
 from pathlib import Path
-
+import psutil
 import flywheel
 
 
-#### Setup logging as per SSE best practices (Thanks Andy!)
-fmt = '%(asctime)s %(levelname)8s %(name)-8s - %(message)s'
-logging.basicConfig(format=fmt)
-log = logging.getLogger('[flywheel/MSOT-mouse-recon]')
-
-supported_afni_opts = OrderedDict()
-supported_afni_opts['copy_anat']=Path
-supported_afni_opts['anat_has_skull']='yn'
-supported_afni_opts['dsets_me_run']=[Path]
-supported_afni_opts['echo_times']=[float]
-supported_afni_opts['reg_echo']=str
-supported_afni_opts['tcat_remove_first_trs']=int
-supported_afni_opts['cost']=str
-supported_afni_opts['tlrc_base']=Path
-supported_afni_opts['tlrc_NL_warp']=bool
-supported_afni_opts['tlrc_no_ss']=bool
-supported_afni_opts['volreg_align_to']=str
-supported_afni_opts['volreg_align_e2a']=bool
-supported_afni_opts['volreg_tlrc_warp']=bool
-supported_afni_opts['mask_epi_anat']='yn'
-supported_afni_opts['combine_method']=str
-supported_afni_opts['combine_opts_tedana']=float
-supported_afni_opts['regress_motion_per_run']=bool
-supported_afni_opts['regress_censor_motion']=float
-supported_afni_opts['regress_censor_outliers']=float
-supported_afni_opts['regress_apply_mot_types']=str
-supported_afni_opts['regress_est_blur_epits']=bool
-
-cost_lookup = {"leastsq":"ls",
-        "mutualinfo":'mi',
-        "corratio_mul":"crM",
-        "norm_mutualinfo":"nmi",
-        "hellinger":"hel",
-        "corratio_add":"crA",
-        "corratio_uns":"crU",
-        "localPcorSigned":"lpc",
-        "localPcorAbs":"lpa",
-        "localPcor+":"lpc+ZZ",
-        "localPcorAbs+":"lpa+ZZ"}
+FORMAT = "[%(filename)s:%(lineno)s - %(funcName)20s() ] %(message)s"
+logging.basicConfig(format=FORMAT)
+log = logging.getLogger('meica')
+log.setLevel('INFO')
 
 
 
@@ -111,7 +75,8 @@ def zipdir(dirPath=None, zipFilePath=None, includeDirInZip=True, deflate=True):
             outFile.writestr(zipInfo, "")
     outFile.close()
 
-def debug_get_meica_data(config, output_directory='/flywheel/v0/output'):
+
+def get_meica_data(context, output_directory='/flywheel/v0/output'):
     """
     For a given input dicom file, grab all of the nifti files from that acquisition.
 
@@ -119,10 +84,10 @@ def debug_get_meica_data(config, output_directory='/flywheel/v0/output'):
     """
 
     # Flywheel Object
-    fw = flywheel.Client(config['inputs']['api_key']['key'])
+    fw = flywheel.Client(context.get_input('api_key')['key'])
 
     # For this acquisition find each nifti file, download it and note its echo time
-    acquisition = fw.get_acquisition(config['inputs'].get('functional').get('hierarchy').get('id'))
+    acquisition = fw.get_acquisition(context.get_input('functional')['hierarchy']['id'])
     nifti_files = [ x for x in acquisition.files
                         if x.type == 'nifti'
                         and "Functional" in x.classification['Intent']
@@ -161,57 +126,6 @@ def debug_get_meica_data(config, output_directory='/flywheel/v0/output'):
 
     return(datasets,tes)
 
-def get_meica_data(context, output_directory='/flywheel/v0/output'):
-    """
-    For a given input dicom file, grab all of the nifti files from that acquisition.
-
-    Return MEICA data which is a sorted list of file objects.
-    """
-
-    # Flywheel Object
-    fw = flywheel.Client(context.get_input('api_key')['key'])
-
-    # For this acquisition find each nifti file, download it and note its echo time
-    acquisition = fw.get_acquisition(context.get_input('functional')['hierarchy']['id'])
-    nifti_files = [ x for x in acquisition.files
-                        if x.type == 'nifti'
-                        and "Functional" in x.classification['Intent']
-                  ]
-    log.info('Found %d Functional NIfTI files in %s' % (len(nifti_files), acquisition.label))
-
-    # Compile meica_data structure
-    meica_data = []
-    repetition_time = ''
-    for n in nifti_files:
-        file_path = os.path.join(output_directory, n.name)
-        log.info('Downloading %s' % (n.name))
-        #fw.download_file_from_acquisition(acquisition.id, n.name, file_path)
-        echo_time = n.info.get('EchoTime')
- 
-        # TODO: Handle case where EchoTime is not here
-        # or classification is not correct
-        # Or not multi echo data
-        # Or if coronavirus attacks 
-
-        meica_data.append({
-                "path": n.name,
-                "te": echo_time*1000 # Convert to ms
-            })
-
-    # Generate prefix
-    sub_code = fw.get_session(acquisition.parents.session).subject.code.strip().replace(' ','')
-    label = acquisition.label.strip().replace(' ','')
-    prefix = '%s_%s' % (sub_code, label)
-    
-    meica_data = sorted(meica_data, key=lambda k: k['te'])
-    datasets = [Path(meica['path']) for meica in meica_data]
-    tes = [meica['te'] for meica in meica_data]
-    
-
-
-    return(datasets,tes)
-
-
 
 def log_system_resources():
     log.info('Logging System Resources\n\n==============================================================================\n')
@@ -232,82 +146,6 @@ def log_system_resources():
     log.info('\n\n==============================================================================\n')
     
     
-def build_afni_proc_call(config):
-    
-    # Starting command (This version of the gear doesn't change these steps
-    command = ['/root/abin/afni_proc.py', '-subj_id', 'data', '-blocks', 'tshift',
-               'align', 'tlrc', 'volreg', 'mask', 'combine', 'blur', 'scale', 'regress']
-    
-    try:
-        for key in supported_afni_opts.keys():
-            kind = supported_afni_opts[key]
-            
-            if isinstance(kind, list):
-                kind = kind[0]
-                
-            if key in config:
-                
-                # a couple odd cases to address
-                if key == 'cost':
-                    data = cost_lookup[config[key]]
-                    append = ['-align_opts_aea', '-{}'.format(key), '{}'.format(data)]
-                    
-                elif key == 'combine_opts_tedana':
-                    data = config['kdaw']
-                    append = ['-{}'.format(key), '--kdaw={}'.format(data)]
-                    
-                # If the expected type is Path
-                elif kind == Path or kind == int or kind == float:
-                    data = config[key]
-                    
-                    if isinstance(data, list):
-                        data = ' '.join([str(item) for item in data])
-                    else:
-                        data = '{data}'.format(data=data)
-                    
-                    append = ['-{}'.format(key)]
-                    append.extend('{}'.format(data).split(' '))
-                    
-                # If the expected type is string
-                elif kind == str:
-                    data = config[key]
-                    
-                    if isinstance(data,list):
-                        data = ' '.join(data)
-
-                    append = ['-{}'.format(key)]
-                    append.extend('{}'.format(data).split(' '))
-                    
-                # If we need the type to be yn (yes/no), the data coming in is a boolean.
-                elif kind == 'yn':
-                    data = config[key]
-                    if data:
-                        data = 'yes'
-                    else:
-                        data = 'no'
-
-                    append = ['-{}'.format(key)]
-                    append.extend('{}'.format(data).split(' '))
-                    
-                elif kind == bool:
-                    data = config[key]
-                    append = ''
-                    if data:
-                        append = ['-{}'.format(key)]
-                    
-                command.extend(append)
-                
-    except Exception as e:
-        log.exception(e)
-        raise e
-    
-    log.info(command)
-    
-    return(command)
-                
-                
-
-
